requirements.txt
aiohttp==3.9.0
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.7
defusedxml==0.8.0rc2
Django==4.1.13
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
dnspython==2.4.2
drf-yasg==1.21.7
frozenlist==1.4.0
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.5.2
MarkupSafe==2.1.3
mongoengine==0.27.0
multidict==6.0.4
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.2
Pillow==10.1.0
pycparser==2.21
PyJWT==2.8.0
pytesseract==0.3.10
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
simplejson==3.19.2
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.2.4
twilio==8.10.2
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
wget==3.2
yarl==1.9.3
